// src/API.ts
export type Product = {
  __typename: "Product",
  id: string,
  image: string,
  images: Array< string >,
  brand: string,
  price: number,
  rating: number,
  numReview: number,
  isFeatured?: boolean | null,
  name: string,
  description: string,
  countInStock?: number | null,
  category?: Category | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  categoryProductsId?: string | null,
};

export type Category = {
  __typename: "Category",
  id: string,
  name: string,
  src: string,
  products?: ModelProductConnection | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};

export type ModelProductConnection = {
  __typename: "ModelProductConnection",
  items:  Array<Product | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};
// ...
export type FavoriteProduct = {
  __typename: "FavoriteProduct",
  id: string,
  userSub: string,
  productID: string,
  product?: Product | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  favoriteProductProductId?: string | null,
};
// ...
export type Message = {
  __typename: "Message",
  id: string,
  image: string,
  productName: string,
  sellerName: string,
  situation: string,
  date: string,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};

export type UpdateMessageInput = {
  id: string,
  image?: string | null,
  productName?: string | null,
  sellerName?: string | null,
  situation?: string | null,
  date?: string | null,
  _version?: number | null,
};

export type DeleteMessageInput = {
  id: string,
  _version?: number | null,
};

export type ModelProductFilterInput = {
  id?: ModelIDInput | null,
  image?: ModelStringInput | null,
  images?: ModelStringInput | null,
  brand?: ModelStringInput | null,
  price?: ModelFloatInput | null,
  rating?: ModelFloatInput | null,
  numReview?: ModelIntInput | null,
  isFeatured?: ModelBooleanInput | null,
  name?: ModelStringInput | null,
  description?: ModelStringInput | null,
  countInStock?: ModelIntInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  and?: Array< ModelProductFilterInput | null > | null,
  or?: Array< ModelProductFilterInput | null > | null,
  not?: ModelProductFilterInput | null,
  _deleted?: ModelBooleanInput | null,
  categoryProductsId?: ModelIDInput | null,
};

export type ModelCategoryFilterInput = {
  id?: ModelIDInput | null,
  name?: ModelStringInput | null,
  src?: ModelStringInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  and?: Array< ModelCategoryFilterInput | null > | null,
  or?: Array< ModelCategoryFilterInput | null > | null,
  not?: ModelCategoryFilterInput | null,
  _deleted?: ModelBooleanInput | null,
};

export type ModelCategoryConnection = {
  __typename: "ModelCategoryConnection",
  items:  Array<Category | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type ModelFavoriteProductFilterInput = {
  id?: ModelIDInput | null,
  userSub?: ModelStringInput | null,
  productID?: ModelIDInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  and?: Array< ModelFavoriteProductFilterInput | null > | null,
  or?: Array< ModelFavoriteProductFilterInput | null > | null,
  not?: ModelFavoriteProductFilterInput | null,
  _deleted?: ModelBooleanInput | null,
  favoriteProductProductId?: ModelIDInput | null,
};

export type ModelFavoriteProductConnection = {
  __typename: "ModelFavoriteProductConnection",
  items:  Array<FavoriteProduct | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};
// ...
export type ModelMessageConnection = {
  __typename: "ModelMessageConnection",
  items:  Array<Message | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};
// ...
export type GetProductQueryVariables = {
  id: string,
};

export type GetProductQuery = {
  getProduct?:  {
    __typename: "Product",
    id: string,
    image: string,
    images: Array< string >,
    brand: string,
    price: number,
    rating: number,
    numReview: number,
    isFeatured?: boolean | null,
    name: string,
    description: string,
    countInStock?: number | null,
    category?:  {
      __typename: "Category",
      id: string,
      name: string,
      src: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    categoryProductsId?: string | null,
  } | null,
};

export type ListProductsQueryVariables = {
  filter?: ModelProductFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListProductsQuery = {
  listProducts?:  {
    __typename: "ModelProductConnection",
    items:  Array< {
      __typename: "Product",
      id: string,
      image: string,
      images: Array< string >,
      brand: string,
      price: number,
      rating: number,
      numReview: number,
      isFeatured?: boolean | null,
      name: string,
      description: string,
      countInStock?: number | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      categoryProductsId?: string | null,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};

export type SyncProductsQueryVariables = {
  filter?: ModelProductFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  lastSync?: number | null,
};

export type SyncProductsQuery = {
  syncProducts?:  {
    __typename: "ModelProductConnection",
    items:  Array< {
      __typename: "Product",
      id: string,
      image: string,
      images: Array< string >,
      brand: string,
      price: number,
      rating: number,
      numReview: number,
      isFeatured?: boolean | null,
      name: string,
      description: string,
      countInStock?: number | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      categoryProductsId?: string | null,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};

export type GetCategoryQueryVariables = {
  id: string,
};

export type GetCategoryQuery = {
  getCategory?:  {
    __typename: "Category",
    id: string,
    name: string,
    src: string,
    products?:  {
      __typename: "ModelProductConnection",
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type ListCategoriesQueryVariables = {
  filter?: ModelCategoryFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListCategoriesQuery = {
  listCategories?:  {
    __typename: "ModelCategoryConnection",
    items:  Array< {
      __typename: "Category",
      id: string,
      name: string,
      src: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};

export type SyncCategoriesQueryVariables = {
  filter?: ModelCategoryFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  lastSync?: number | null,
};

export type SyncCategoriesQuery = {
  syncCategories?:  {
    __typename: "ModelCategoryConnection",
    items:  Array< {
      __typename: "Category",
      id: string,
      name: string,
      src: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};

export type GetFavoriteProductQueryVariables = {
  id: string,
};

export type GetFavoriteProductQuery = {
  getFavoriteProduct?:  {
    __typename: "FavoriteProduct",
    id: string,
    userSub: string,
    productID: string,
    product?:  {
      __typename: "Product",
      id: string,
      image: string,
      images: Array< string >,
      brand: string,
      price: number,
      rating: number,
      numReview: number,
      isFeatured?: boolean | null,
      name: string,
      description: string,
      countInStock?: number | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      categoryProductsId?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    favoriteProductProductId?: string | null,
  } | null,
};

export type ListFavoriteProductsQueryVariables = {
  filter?: ModelFavoriteProductFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListFavoriteProductsQuery = {
  listFavoriteProducts?:  {
    __typename: "ModelFavoriteProductConnection",
    items:  Array< {
      __typename: "FavoriteProduct",
      id: string,
      userSub: string,
      productID: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      favoriteProductProductId?: string | null,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};

export type SyncFavoriteProductsQueryVariables = {
  filter?: ModelFavoriteProductFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  lastSync?: number | null,
};

export type SyncFavoriteProductsQuery = {
  syncFavoriteProducts?:  {
    __typename: "ModelFavoriteProductConnection",
    items:  Array< {
      __typename: "FavoriteProduct",
      id: string,
      userSub: string,
      productID: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      favoriteProductProductId?: string | null,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};

export type GetMessageQueryVariables = {
  id: string,
};

export type GetMessageQuery = {
  getMessage?:  {
    __typename: "Message",
    id: string,
    image: string,
    productName: string,
    sellerName: string,
    situation: string,
    date: string,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type ListMessagesQueryVariables = {
  filter?: ModelMessageFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListMessagesQuery = {
  listMessages?:  {
    __typename: "ModelMessageConnection",
    items:  Array< {
      __typename: "Message",
      id: string,
      image: string,
      productName: string,
      sellerName: string,
      situation: string,
      date: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};

export type SyncMessagesQueryVariables = {
  filter?: ModelMessageFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  lastSync?: number | null,
};

export type SyncMessagesQuery = {
  syncMessages?:  {
    __typename: "ModelMessageConnection",
    items:  Array< {
      __typename: "Message",
      id: string,
      image: string,
      productName: string,
      sellerName: string,
      situation: string,
      date: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null >,
    nextToken?: string | null,
    startedAt?: number | null,
  } | null,
};